Work around truncated macho files (#52)
* Work around truncated macho files

This avoids panicking when iterating over sections if an improperly
truncated macho files is encountered.  This is happening for files
that are provided by customers we have no control over.  These files
currently read fine with macOS tools presumably because those sections
are not ever used (we only care about DWARF data).

* Emit a warning when a section is empty