elf: do not fail entire parse for bad interpreters, fixes #192
Some tools leave a PT_INTERP Phdr in a bad way when manipulating the ELF file (for example,
https://bugzilla.redhat.com/show_bug.cgi?id=1770513 where splitting debug
info into a separate file leaves a bogus PT_INTERP header). When using
`gimli` with `object` to parse DWARF info, the PT_INTERP is irrelevent
so failing to load the file because of it causes loss of functionality.