mach: Support archive entries in multi-arch containers
Multi-arch containers can be made up of archives or Mach-O
binaries. This adds support for archives. It is a breaking change
because previously the `MachO` struct was returned and now we're
returning a new enum: `SingleArch`.

This required some refactoring of the `lib.rs` file to share the
required functions and data structures for parsing the hint at the top
of files. `take_hint_bytes`, `Hint` and `HintData` dont't require any
special features but I think were inside the `if_everything!` because
that's the only case they were used in. I did expand the API by making
`take_byte_hints` public, this was mainly because I thought it was
better solution than trying to maintain the various combinations of
features required to stop the compiler warning about the function
being unused. It's also a function that may be useful for goblin users.

resolves https://github.com/m4b/goblin/issues/320