mach: Add tests for parsing fat binaries
This tests parsing fat binaries made up of Mach-O binaries and
archives. I've checked in the binaries to make testing easier as
they're quite small (both are built from the hello_world.c file in the
same directory). Above the tests themselves are instructions for how
to compile the binaries they use.