pe(writer): introduce a meta Writer structure
Writing new structures from an existing PE is non trivial because PE structures
may require a complete rewrite of all the structures and recompute of the offsets for space
reasons.

We sidestep the difficulty by having a meta Writer structure which tracks the information required
to re-render an existing PE information + extra informations like new sections.