pe: write support for PE binaries (#361)
* strtab: offer `len` method

* pe(header): introduce full DOS header and DOS stub

Required to rematerialize a whole PE, otherwise, it is a corrupted PE.

* pe(section_table): add `data` function to retrieve the section data safely

It now returns a Cow as we need to transfer ownership whenever we have
virtual size > size of raw data, which forces us to pad with zeroes
according to the PE specification.

* pe(utils): add a zero-extension padding helper

It can be used to zero-extend any buffer to make its length
aligned on some boundary, e.g. quadwords, for attribute certificates.

It is designed to be "zero cost" if you have no alignment to do, which
will result in no-allocation!

* pe: implement write support for AttributeCertificate

* pe: implement write support for DataDirectories

* pe: implement write support for Header

It enables a lossless conversion from 64 bits to 32 bits headers
except on platforms… that are strange.

* pe: implement write support for SymbolTable

* pe: implement write support for data directories

This tucks `DataDirectory` with a Deref-style and expose `offset` for internal consumers (this crate)
and offers write support for data directories.

* pe: implement write support for SectionTable

All that is needed is to import the allocation `Vec`.

* pe: implement write support

* examples(pe): add a identity rewrite PE binary

This is a trivial rewriter that performs the identity transformation.

* pe: debug assert that no overlapping sections are written

To further improve the debuggability, a non-overlapping sanity check is added as a
debug_assert!.

* pe(coff): COFF symbols and COFF strings are deprecated

Previously, I (we?) thought that COFF symbols/strings were always there.

In fact, this is not the case and they are deprecated
according to https://learn.microsoft.com/en-us/windows/win32/debug/pe-format#coff-file-header-object-and-image

> The file offset of the COFF symbol table, or zero
> if no COFF symbol table is present. This value should be zero for an image
> because COFF debugging information is deprecated.

* pe(write): factor into `write_certificates`

We had the write certificate logic in the `TryIntoCtx`, but it makes
sense to have it separated to let consumers call it manually.