Fix issue 370
The issue gets triggered, because qemu/virsh generated elf dump contains strtab
at the end of the dump file. Meanwhile, dumps are large, and thus, it is
infeasible to provide the entire dump file to the parser. This PR relaxes the
requirements of the parser and returns default strtab, if the initial size
check fails.